pyproject.toml
[project]
name = "argo-ai"
version = "0.2.0"
description = "Agent-based Reasoning, Governance, and Orchestration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.76.0",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]

[project.scripts]
argo = "argo.cli:main"

[dependency-groups]
dev = [
    "dotenv>=0.9.9",
    "googlesearch-python>=1.3.0",
    "markitdown>=0.1.1",
    "pytest>=8.3.5",
    "python-dotenv[cli]>=1.1.0",
    "ruff>=0.11.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["argo"]
